Updated to generate XML loan response
Loan response was updated to produce XBRL. But it is reverted as per the discussion. Now we have the ledger response in XBRL and loan response in XML.

Change in loan response:
----------------------------
Namespaces from the <LoanSummary> node is removed. When we have a schema for loan summary, we should refer that schema in the loan response XML.